"v2 오류수정"
