+++ app/src/main/java/com/example/sundo_project_app/location/MapActivity.java
@@ -2,6 +2,7 @@ package com.example.sundo_project_app.location;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.widget.Button;
 import android.widget.Toast;
 
@@ -9,6 +10,8 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.sundo_project_app.R;
+import com.example.sundo_project_app.evaluation.EvaluationActivity;
+import com.example.sundo_project_app.evaluation.EvaluationDialogFragment;
 import com.naver.maps.map.MapFragment;
 import com.naver.maps.map.NaverMap;
 import com.naver.maps.map.overlay.Marker;
@@ -18,6 +21,8 @@ public class MapActivity extends AppCompatActivity {
     private NaverMap naverMap;
     private boolean isMarkerEnabled = false; // 마커 추가 모드 상태
     private Button coordinateSelectButton;
+    private Button btnShowDialog;
+    private Button btnShowList;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -45,6 +50,25 @@ public class MapActivity extends AppCompatActivity {
         // 좌표 선택 버튼 초기화
         coordinateSelectButton = findViewById(R.id.coordinateSelect);
         coordinateSelectButton.setOnClickListener(v -> toggleMarkerMode());
+
+        btnShowDialog = findViewById(R.id.btnShowDialog);
+        btnShowList = findViewById(R.id.enteredPoint);
+
+        btnShowDialog.setOnClickListener(v -> {
+            Log.d("EvaluationFindAllActivity", "평가리스트 버튼 클릭됨");
+            showEvaluationDialog();
+        });
+
+        btnShowList.setOnClickListener(v -> {
+            Log.d("btnShowList", "평가입력 버튼 클릭됨");
+            Intent intent = new Intent(MapActivity.this, EvaluationActivity.class);
+            startActivity(intent);
+        });
+    }
+
+    private void showEvaluationDialog() {
+        EvaluationDialogFragment dialog = new EvaluationDialogFragment();
+        dialog.show(getSupportFragmentManager(), "EvaluationDialog");
     }
 
     private void onMapReady(@NonNull NaverMap naverMap) {
